# Fixing perf_client to generate valid shared memory region names