### Before
1. ../clients/perf_client -m inception_graphdef -b 1 --shared-memory=system
[ 0] INVALID_ARG - unable to get shared memory descriptor for shared-memory key '/InceptionV3/Predictions/Softmax'

2. ../clients/perf_client -m inception_graphdef -b 1 --shared-memory=cuda        
[ 0] INTERNAL - sharedmemorycontrol request did not return status

### After
 1. ../clients/perf_client -m inception_graphdef -b 1 --shared-memory=system
*** Measurement Settings ***
  Batch size: 1
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Client: 
    Request count: 412
    Throughput: 82.4 infer/sec
    Avg latency: 12121 usec (standard deviation 1161 usec)
    p50 latency: 12476 usec
    p90 latency: 13208 usec
    p95 latency: 13465 usec
    p99 latency: 14448 usec
    Avg HTTP time: 12128 usec (send/recv 2 usec + response wait 12126 usec)
  Server: 
    Request count: 494
    Avg request latency: 11266 usec (overhead 11 usec + queue 40 usec + compute 11215 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, throughput: 82.4 infer/sec, latency 12121 usec

2.  ../clients/perf_client -m inception_graphdef -b 1 --shared-memory=cuda  
*** Measurement Settings ***
  Batch size: 1
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Client: 
    Request count: 396
    Throughput: 79.2 infer/sec
    Avg latency: 12608 usec (standard deviation 977 usec)
    p50 latency: 12872 usec
    p90 latency: 13549 usec
    p95 latency: 13731 usec
    p99 latency: 14265 usec
    Avg HTTP time: 12552 usec (send/recv 2 usec + response wait 12550 usec)
  Server: 
    Request count: 477
    Avg request latency: 11686 usec (overhead 12 usec + queue 43 usec + compute 11631 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, throughput: 79.2 infer/sec, latency 12608 usec